added anxieties and values

+++ math.c
@@ -20,7 +20,7 @@ float ve_dot(struct vect v1, struct vect v2){
 
 float ve_angle(float sqr1, float sqr2, float dprod){
     float multSqr = sqr1 * sqr2;
-    acos(dprod * invSqr(multSqr));
+    acosf(dprod * invSqr(multSqr));
 }
 /*******************************************************/
